Remove TouchState from Enemy View and Presenter

+++ MorskoiBoi/src/main/java/com/ivygames/morskoiboi/renderer/BaseBoardRenderer.java
@@ -77,7 +77,7 @@ public class BaseBoardRenderer {
         canvas.drawRect(aiming.vertical, paint);
     }
 
-    public void drawDebug(@NonNull Canvas canvas, int x, int y) {
+    public void drawDebug(@NonNull Canvas canvas, float x, float y) {
         canvas.drawCircle(x, y, 5, debug_paint);
     }
